Move type-checking logic in StaticConst to RedundantStaticLifetime.
